create AddList component and move related code there
